I so much goddamn hate async

+++ chrome/content/zotero-better-bibtex/db.coffee
@@ -1,29 +1,67 @@
 Components.utils.import('resource://gre/modules/Services.jsm')
 
-Zotero.BetterBibTeX.DBStore = new class
-  constructor: ->
-    @store = new Zotero.DBConnection('betterbibtex-lokijs')
-    @store.query('CREATE TABLE IF NOT EXISTS lokijs (name PRIMARY KEY, data)')
+if Zotero.BetterBibTeX.Async
+  Zotero.BetterBibTeX.debug('DBStore: Async')
+  Zotero.BetterBibTeX.DBStore = new class
+    constructor: ->
+      @store = new Zotero.DBConnection('betterbibtex-lokijs')
+      Zotero.BetterBibTeX.debug('DBStore: Async.init')
+      #@store = new Zotero.BetterBibTeX.SQLite('betterbibtex-lokijs')
+      #@store.query('CREATE TABLE IF NOT EXISTS lokijs (name PRIMARY KEY, data)')
 
-  saveDatabase: (name, serialized, callback) ->
-    if !Zotero.initialized || Zotero.isConnector
-      Zotero.BetterBibTeX.flash('Zotero is in connector mode -- not saving database!')
-    else
-      Zotero.BetterBibTeX.debug("Saving database #{name}")
-      @store.query("INSERT OR REPLACE INTO lokijs (name, data) VALUES (?, ?)", [name, serialized])
-    callback()
-    return
+    saveDatabase: (name, serialized, callback) ->
+      if !Zotero.initialized || Zotero.isConnector
+        Zotero.BetterBibTeX.flash('Zotero is in connector mode -- not saving database!')
+        return
 
-  loadDatabase: (name, callback) ->
-    file = Zotero.BetterBibTeX.createFile(name)
-    if file.exists()
-      Zotero.BetterBibTeX.debug('DB.loadDatabase:', {name, file: file.path})
-      callback(Zotero.File.getContents(file))
-      file.remove(null) if file.exists()
+      Zotero.BetterBibTeX.debug('DBStore: Async.save')
+
+      Zotero.Promise.coroutine(->
+        try
+          yield store.queryAsync('CREATE TABLE IF NOT EXISTS lokijs (name PRIMARY KEY, data)')
+          yield store.queryAsync('INSERT OR REPLACE INTO lokijs (name, data) VALUES (?, ?)', [name, serialized])
+          callback()
+        catch err
+          callback(err)
+      )()
+
+    loadDatabase: (name, callback) ->
+      Zotero.BetterBibTeX.debug('DBStore: Async.load')
+      Zotero.Promise.coroutine(=>
+        try
+          yield @store.queryAsync('CREATE TABLE IF NOT EXISTS lokijs (name PRIMARY KEY, data)')
+          data = yield @store.valueQueryAsync('SELECT data FROM lokijs WHERE name=?', [name])
+          callback(data || null)
+        catch err
+          callback(err)
+      )()
+
+else
+  Zotero.BetterBibTeX.debug('DBStore: Sync')
+  Zotero.BetterBibTeX.DBStore = new class
+    constructor: ->
+      @store = new Zotero.DBConnection('betterbibtex-lokijs')
+      @store.query('CREATE TABLE IF NOT EXISTS lokijs (name PRIMARY KEY, data)')
+
+    saveDatabase: (name, serialized, callback) ->
+      if !Zotero.initialized || Zotero.isConnector
+        Zotero.BetterBibTeX.flash('Zotero is in connector mode -- not saving database!')
+      else
+        Zotero.BetterBibTeX.debug("Saving database #{name}")
+        @store.query("INSERT OR REPLACE INTO lokijs (name, data) VALUES (?, ?)", [name, serialized])
+      callback()
       return
 
-    callback(@store.valueQuery("SELECT data FROM lokijs WHERE name=?", [name]) || null)
-    return
+    loadDatabase: (name, callback) ->
+      file = Zotero.BetterBibTeX.createFile(name)
+      if file.exists()
+        Zotero.BetterBibTeX.debug('DB.loadDatabase:', {name, file: file.path})
+        callback(Zotero.File.getContents(file))
+        file.remove(null) if file.exists()
+        return
+
+      callback(@store.valueQuery("SELECT data FROM lokijs WHERE name=?", [name]) || null)
+      return
 
 Zotero.BetterBibTeX.DB = new class
   cacheExpiry: Date.now() - (1000 * 60 * 60 * 24 * 30)
@@ -114,7 +152,7 @@ Zotero.BetterBibTeX.DB = new class
 
     @upgradeNeeded = {}
     freshInstall = true
-    for k, v of { Zotero: ZOTERO_CONFIG.VERSION, BetterBibTeX: Zotero.BetterBibTeX.release, storage: Zotero.getZoteroDirectory().path }
+    for k, v of { Zotero: Zotero.BetterBibTeX.zoteroRelease, BetterBibTeX: Zotero.BetterBibTeX.release, storage: Zotero.getZoteroDirectory().path }
       freshInstall = false if @metadata[k]
       continue if @metadata[k] == v
       @upgradeNeeded[k] = v
@@ -240,7 +278,7 @@ Zotero.BetterBibTeX.DB = new class
     Zotero.debug('DB.initialize: ready')
 
   purge: ->
-    itemIDs = Zotero.DB.columnQuery('select itemID from items except select itemID from deletedItems')
+    itemIDs = Zotero.BetterBibTeX.SQLite::columnQuery('select itemID from items except select itemID from deletedItems')
     itemIDs = (parseInt(id) for id in itemIDs)
     @keys.removeWhere((o) -> o.itemID not in itemIDs)
     @cache.removeWhere((o) -> o.itemID not in itemIDs)
@@ -286,141 +324,3 @@ Zotero.BetterBibTeX.DB = new class
           throw(err)
       )
       @db.main.autosaveClearFlags()
-
-  SQLite:
-    parseTable: (name) ->
-      name = name.split('.')
-      switch name.length
-        when 1
-          schema = ''
-          name = name[0]
-        when 2
-          schema = name[0] + '.'
-          name = name[1]
-      name = name.slice(1, -1) if name[0] == '"'
-      return {schema: schema, name: name}
-
-    table_info: (table) ->
-      table = @parseTable(table)
-      statement = Zotero.DB.getStatement("pragma #{table.schema}table_info(\"#{table.name}\")", null, true)
-
-      fields = (statement.getColumnName(i).toLowerCase() for i in [0...statement.columnCount])
-
-      columns = {}
-      while statement.executeStep()
-        values = (Zotero.DB._getTypedValue(statement, i) for i in [0...statement.columnCount])
-        column = {}
-        for name, i in fields
-          column[name] = values[i]
-        columns[column.name] = column
-      statement.finalize()
-
-      return columns
-
-    columnNames: (table) ->
-      return Object.keys(@table_info(table))
-
-    tableExists: (name) ->
-      table = @parseTable(name)
-      return (Zotero.DB.valueQuery("SELECT count(*) FROM #{table.schema}sqlite_master WHERE type='table' and name=?", [table.name]) != 0)
-
-    Set: (values) -> '(' + ('' + v for v in values).join(', ') + ')'
-
-    migrate: ->
-      db = Zotero.BetterBibTeX.createFile('serialized-items.json')
-      db.remove(null) if db.exists()
-
-      db = Zotero.BetterBibTeX.createFile('db.json.bak')
-      db.remove(null) if db.exists()
-
-      db = Zotero.BetterBibTeX.createFile('cache.json.bak')
-      db.remove(null) if db.exists()
-
-      db = Zotero.getZoteroDatabase('betterbibtexcache')
-      db.remove(true) if db.exists()
-
-      db = Zotero.BetterBibTeX.createFile('better-bibtex-serialized-items.json')
-      db.remove(true) if db.exists()
-
-      db = Zotero.getZoteroDatabase('..', 'betterbibtex.sqlite.bak')
-      db.remove(true) if db.exists()
-
-      db = Zotero.getZoteroDatabase('betterbibtex')
-      return false unless db.exists()
-
-      Zotero.BetterBibTeX.flash('Better BibTeX: updating database', 'Updating database, this could take a while')
-
-      Zotero.DB.query('ATTACH ? AS betterbibtexmigration', [db.path])
-
-      if @tableExists('betterbibtexmigration.autoexport') && !@table_info('betterbibtexmigration.autoexport').context
-        Zotero.BetterBibTeX.debug('DB.migrate: autoexport')
-        Zotero.BetterBibTeX.DB.autoexport.removeDataOnly()
-
-        if @table_info('betterbibtexmigration.autoexport').collection
-          Zotero.DB.query("update betterbibtexmigration.autoexport set collection = (select 'library:' || libraryID from groups where 'group:' || groupID = collection) where collection like 'group:%'")
-          Zotero.DB.query("update betterbibtexmigration.autoexport set collection = 'collection:' || collection where collection <> 'library' and collection not like '%:%'")
-
-        migrated = 0
-        for row in Zotero.DB.query('select * from betterbibtexmigration.autoexport')
-          migrated += 1
-          Zotero.BetterBibTeX.DB.autoexport.insert({
-            collection: row.collection
-            path: row.path
-            exportCharset: row.exportCharset
-            exportNotes: (row.exportNotes == 'true')
-            translatorID: row.translatorID
-            useJournalAbbreviation: (row.useJournalAbbreviation == 'true')
-            status: 'pending'
-          })
-        Zotero.BetterBibTeX.debug('DB.migrate: autoexport=', migrated)
-
-      if @tableExists('betterbibtexmigration.cache')
-        Zotero.BetterBibTeX.debug('DB.migrate: cache')
-        Zotero.BetterBibTeX.DB.cache.removeDataOnly()
-
-        migrated = 0
-        for row in Zotero.DB.query('select * from betterbibtexmigration.cache')
-          migrated += 1
-          Zotero.BetterBibTeX.DB.cache.insert({
-            itemID: parseInt(row.itemID)
-            exportCharset: row.exportCharset
-            exportNotes: (row.exportNotes == 'true')
-            translatorID: row.translatorID
-            useJournalAbbreviation: (row.useJournalAbbreviation == 'true')
-            citekey: row.citekey
-            bibtex: row.bibtex
-            accessed: Date.now()
-          })
-
-        Zotero.BetterBibTeX.debug('DB.migrate: cache=', migrated)
-
-      if @tableExists('betterbibtexmigration.keys')
-        Zotero.BetterBibTeX.debug('DB.migrate: keys')
-        Zotero.BetterBibTeX.DB.keys.removeDataOnly()
-        pinned = @table_info('betterbibtexmigration.autoexport').pinned
-
-        migrated = 0
-        for row in Zotero.DB.query('select k.*, i.libraryID from betterbibtexmigration.keys k join items i on k.itemID = i.itemID')
-          continue if pinned && row.pinned != 1
-          migrated += 1
-
-          row.citekeyFormat = null unless row.citekeyFormat
-
-          Zotero.BetterBibTeX.DB.keys.insert({
-            itemID: parseInt(row.itemID)
-            citekey: row.citekey
-            citekeyFormat: row.citekeyFormat
-            libraryID: row.libraryID
-          })
-        Zotero.BetterBibTeX.debug('DB.migrate: keys=', migrated)
-
-      Zotero.DB.query('DETACH betterbibtexmigration')
-
-      db.moveTo(null, 'betterbibtex.sqlite.bak')
-
-      Zotero.BetterBibTeX.DB.save('force')
-
-      Zotero.BetterBibTeX.flash('Better BibTeX: database updated', 'Database update finished')
-      Zotero.BetterBibTeX.flash('Better BibTeX: cache has been reset', 'Cache has been reset due to a version upgrade. First exports after upgrade will be slower than usual')
-
-      return true
